feat: 그룹api중간저장
