+++ Replendar/src/components/CommuComponents/modalContents/makeGroup.tsx
@@ -2,9 +2,45 @@ import GrayPlusIcon from '../../../assets/images/GrayPlusIcon.svg';
 import { PlusFriendsButton } from '../../../modal/AddTaskModal';
 import useModalStore from '../../../store/modalStore';
 import styled from 'styled-components';
+import { createGroup } from '../../../apis/commuApi';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useState } from 'react';
+import SelectFriendsModal from '../../../modal/SelectFriendsModal';
+import useGetData from '../../../hooks/useGetData';
+import useFriendsStore from '../../../store/useFriendStore';
 
 const MakeGroup: React.FC = () => {
   const { closeModal } = useModalStore();
+
+  const { isFriendModalOpen, openFriendModal } = useFriendsStore();
+
+  const [groupName, setGroupName] = useState('');
+
+  const queryClient = useQueryClient();
+
+  const mutation = useMutation({
+    mutationFn: (groupName: string) => createGroup(groupName),
+    onSuccess: () => {
+      alert('그룹이 성공적으로 생성되었습니다.');
+      closeModal();
+
+      queryClient.invalidateQueries({
+        queryKey: [`/api/friend-groups`],
+      });
+    },
+    onError: (error: Error) => {
+      alert('그룹을 생성하는 데 실패했습니다.');
+      console.error(error);
+    },
+  });
+  const handleCreateGroup = () => {
+    if (!groupName.trim()) {
+      alert('그룹 이름을 입력하세요.');
+      return;
+    }
+    mutation.mutate(groupName);
+  };
+
   return (
     <Container>
       <h1>그룹 만들기</h1>
@@ -12,17 +48,22 @@ const MakeGroup: React.FC = () => {
         <div>
           {' '}
           <label>그룹명</label>
-          <TransparentInput placeholder="그룹 이름을 입력하세요"></TransparentInput>
+          <TransparentInput
+            placeholder="그룹 이름을 입력하세요"
+            value={groupName}
+            onChange={(e) => setGroupName(e.target.value)}
+          ></TransparentInput>
         </div>
 
         <label>그룹에 넣을 친구</label>
       </GapDiv>
-      <PlusFriendsButton>
+      <PlusFriendsButton onClick={openFriendModal}>
         <img src={GrayPlusIcon} alt="Gray Plus Icon" />
         추가
       </PlusFriendsButton>
+      {isFriendModalOpen && <SelectFriendsModal />}
       <FlexEndGroup>
-        <PlusFriendsButton onClick={closeModal}>완료</PlusFriendsButton>
+        <PlusFriendsButton onClick={handleCreateGroup}>완료</PlusFriendsButton>
       </FlexEndGroup>
     </Container>
   );
