+++ Replendar/src/types.ts
@@ -22,3 +22,9 @@ export interface IFriendList {
   ongoingAssignments: number;
   buddyStatus: 'YES' | 'NO';
 }
+
+export interface IGroupList {
+  groupId: number; // 그룹 ID
+  groupName: string; // 그룹 이름
+  friends: IFriendList[]; // 그룹에 속한 친구들의 배열
+}
